Correct SubjectNamingInfo ClaimType
The highlighted SubjectNamingInfo element below, should read <SubjectNamingInfo ClaimType="userPrincipalName " /> since we only want to reference claims that are defined in the policy schema. By default the claim that is used as the subject name will be excluded from the claims output in the token since it is output as the subject.  However an additional attribute has been added to the SubjectNamingInfo element (should got out in deployment for week 1 - July 2017) called IncludeAsClaim which takes true/false, that allows a policy author to override the default behavior and include in the claim as an output claim in addition to being the subject claim.

OLD <SubjectNamingInfo ClaimType="http://schemas.xmlsoap.org/ws/2005/05/identity/claims/upn" />
 NEW <SubjectNamingInfo ClaimType="userPrincipalName" />