A description of how to implement invite flows.
The WingTip demo at aka.ms/aadb2cdemo now includes an example of how to invite a user in two modalities.  For the demo, the user invites herself, but the same functionality is available in other ways.  In this mode, the user does not exist in the directory, until he/she accepts the invitation and the completes a registration

It is also possible for the admin to create a user and then invite her.  To try this, you must have access to the admin portion of the demo at https://wingtiptoysb2c.azurewebsites.net .  

All code is already incorporate in the general wingtiptoys sample.

Thanks to Chris Padgett from Kloud for the dev. work here.