Update Sample now includes treatment for force_change password.
This sample  uses Signup Signin with AAD B2C using custom policies.  It includes the correct default responses when a Conditional Access policy has been defined in the tenant for Block, MFA, and now includes chg_password.  This claim value now results in the user having to 1. MFA, and then 2. provide a new password. It is triggered by a Conditional Access policy, or by a tenant admin action